Fix legacy data in current_contributing_question of submitted_questions
This is a script that should most likely be run in production and then import the data. Although, it will only take about 2 minutes to run in development, and less than 1 minute in production. The way to run this is:  ruby ./script/adjust_contributing_id.rb --environment 'development' > your_out_file_if_desired.out.  If running in production, can subsitute 'production' for 'development' or skip the --environment parm entirely as 'production' is the default. 

     This fixes legacy questions where the current_contributing_id was either the converted FAQ or the actual current contributing question, and data after 8/12/2009 uses an entirely different scheme where the current_contributing_question is an id of a search_question whose foreignid actually is the relevant contributing question.