Fixed Homage Issue in Content Checklist for Preview
Fixed the view bug with this to keep errors from occurring for now, but
the underlying data sync issue still remains. In the
farm_safety_and_health case, the homage for the community is missing,
but there are still pages tagged homage for it. Will log a separate
issue for this.